Added numpy code to test the EBK terms. - Added compute_ebk_jones() to TestSharedData. - Add some asserts into compute_bk_jones() - Fixed the some E term computation in the GPU code. - set_cos3_constant() allows the specification of the cos^3 constant in the E term. - Changed some sqrt's to sqrtf's in the float kernels.

+++ RimeBKFloat.py
@@ -77,10 +77,10 @@ void rime_jones_BK_float(
     __syncthreads();
 
     // Calculate the n term first
-    // n = sqrt(1.0 - l*l - m*m) - 1.0
+    // n = sqrtf(1.0 - l*l - m*m) - 1.0
     float phase = 1.0 - l[threadIdx.x]*l[threadIdx.x];
     phase -= m[threadIdx.x]*m[threadIdx.x];
-    phase = sqrt(phase) - 1.0;
+    phase = sqrtf(phase) - 1.0;
     // TODO: remove this superfluous variable
     // It only exists for debugging purposes
     // float n = phase;
